Add non-ambiguous task name shortcuts
This a first proof-of-concept implementation to allow specifying tasks
on the command line using non ambiguous shortcuts. A shortcut is
computed for each task, splitting its name on underscores and
concatenating the initial letter of each word (for example, the task
name "say_hello_world" would be abbreviated by "shw"): if there are no
other names starting with that, the shortcut is accepted.